Move parser.py to _internal
