+++ certbot-nginx/docs/api/parser.rst
@@ -1,5 +1,5 @@
-:mod:`certbot_nginx.parser`
+:mod:`certbot_nginx._internal.parser`
 -------------------------------
 
-.. automodule:: certbot_nginx.parser
+.. automodule:: certbot_nginx._internal.parser
    :members:
